tech-studio-projects/other-projects/multi-product-recommendation-system/python/recommendation_engine_service/requirements.txt
Django==4.1.5
celery==5.2.7
pytz==2022.7
redis==4.4.0
djangorestframework==3.14.0
